refactor(models): change BigInt field by Decimal

+++ services/wallets/src/domain/entities/transaction/index.ts
@@ -1,4 +1,4 @@
-import { Wallet } from "../wallet";
+import type { Decimal } from "@prisma/client/runtime/client";
 
 export type TRANSACTION_TYPE = "CREDIT" | "DEBIT";
 export type TRANSACTION_STATUS = "PENDING" | "CONFIRMED" | "FAILED";
@@ -7,7 +7,7 @@ interface TransactionProps {
   id: string;
   type: TRANSACTION_TYPE;
   status: TRANSACTION_STATUS;
-  amount: bigint;
+  amount: Decimal;
   createdAt: Date;
   walletId?: string | null;
 }
@@ -16,7 +16,7 @@ export class Transaction implements TransactionProps {
   id: string;
   type: TRANSACTION_TYPE;
   status: TRANSACTION_STATUS;
-  amount: bigint;
+  amount: Decimal;
   createdAt: Date;
   walletId?: string | null;
 }
